+++ src/main/java/com/dapanda/product/dto/MobileDataSummary.java
@@ -1,5 +1,6 @@
 package com.dapanda.product.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,12 +9,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class MobileDataSummary extends ProductSummary {
 
-	private float remainAmount;
+	private BigDecimal remainAmount;
 	private int pricePer100MB;
 	private boolean splitType;
 	private LocalDateTime updatedAt;
 
-	public MobileDataSummary(Long id, int price, Long itemId, String memberName, float remainAmount,
+	public MobileDataSummary(Long id, int price, Long itemId, String memberName,
+			BigDecimal remainAmount,
 			int pricePer100MB, boolean splitType, LocalDateTime updatedAt) {
 
 		super(id, price, itemId, memberName);
