+++ src/main/java/com/dapanda/product/dto/response/MobileDataInfoResponse.java
@@ -1,5 +1,6 @@
 package com.dapanda.product.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -13,7 +14,7 @@ public class MobileDataInfoResponse {
 	private int price;
 	private Long memberId;
 	private String memberName;
-	private float remainAmount;
+	private BigDecimal remainAmount;
 	private int pricePer100MB;
 	private float averageRate;
 	private int reviewCount;
